+++ src/main/java/com/wolApi/wolApi/api/bibleVerses/BibleVerseViews.java
@@ -1,8 +1,8 @@
-package com.wolApi.wolApi.api.getVerse;
+package com.wolApi.wolApi.api.bibleVerses;
 
-import com.wolApi.wolApi.api.getVerse.dtos.Verse;
-import com.wolApi.wolApi.api.getVerse.dtos.VerseList;
-import com.wolApi.wolApi.api.getVerse.services.BibleVerseService;
+import com.wolApi.wolApi.api.bibleVerses.dtos.Verse;
+import com.wolApi.wolApi.api.bibleVerses.dtos.VerseList;
+import com.wolApi.wolApi.api.bibleVerses.services.BibleVerseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
