+++ src/main/java/com/wolApi/wolApi/api/bibleVerses/dtos/Verse.java
@@ -1,0 +1,3 @@
+package com.wolApi.wolApi.api.bibleVerses.dtos;
+
+public record Verse(String data) {}
