--- src/main/java/com/wolApi/wolApi/api/getVerse/dtos/Verse.java
@@ -1,3 +1,0 @@
-package com.wolApi.wolApi.api.getVerse.dtos;
-
-public record Verse(String data) {}
